Add remove button
Add remove button

+++ Project_1AFT/CustomerInfo.vb
@@ -1,6 +1,14 @@
 ﻿Imports System.Data
 Imports System.Data.SqlClient
 Public Class CustomerInfo
+    Public ValCheck As Integer = 0
+    Private Sub ValidationCheck()
+        If TextBox1.Text Is "" Or TextBox2.Text Is "" Or TextBox3.Text Is "" Or TextBox4.Text Is "" Or TextBox4.Text Is "" Then
+            ValCheck = 1
+        Else
+            ValCheck = 0
+        End If
+    End Sub
     Private Sub TextBox4_KeyPress(ByVal sender As System.Object, ByVal e As System.Windows.Forms.KeyPressEventArgs) Handles TextBox4.KeyPress
 
         If Not Char.IsDigit(e.KeyChar) And Not Char.IsControl(e.KeyChar) Then
@@ -63,41 +71,51 @@ Public Class CustomerInfo
     End Sub
 
     Private Sub Button3_Click(sender As Object, e As EventArgs) Handles Button3.Click
-
-        Try
-            Dim con As New SqlConnection(DBconfig)
-            Dim command As New SqlCommand("Update aft_customer
+        ValidationCheck()
+        If ValCheck = 0 Then
+            Try
+                Dim con As New SqlConnection(DBconfig)
+                Dim command As New SqlCommand("Update aft_customer
             SET [cname] = '" + TextBox2.Text + "'
             ,[cemail] = '" + TextBox3.Text + "'
             ,[cphone] = '" + TextBox4.Text + "'
             ,[caddress] = '" + TextBox5.Text + "'
             WHERE icnumber=" + TextBox1.Text + "", con)
-            con.Open()
-            command.ExecuteNonQuery()
-            MessageBox.Show("Customer Info Updated")
-            con.Close()
-        Catch ex As Exception
-            MessageBox.Show("Exception: {0}", ex.Message)
-            MessageBox.Show("We found an error, Please contact with super admin.")
-        End Try
+                con.Open()
+                command.ExecuteNonQuery()
+                MessageBox.Show("Customer Info Updated")
+                con.Close()
+            Catch ex As Exception
+                MessageBox.Show("Exception: {0}", ex.Message)
+                MessageBox.Show("We found an error, Please contact with super admin.")
+
+            End Try
+        Else
+            MessageBox.Show("All information is required.")
+        End If
 
     End Sub
 
     Private Sub Button4_Click(sender As Object, e As EventArgs) Handles Button4.Click
-        Try
-            Dim con As New SqlConnection(DBconfig)
-            Dim command As New SqlCommand("DELETE FROM aft_customer
+        If TextBox1.Text IsNot "" Then
+
+            Try
+                Dim con As New SqlConnection(DBconfig)
+                Dim command As New SqlCommand("DELETE FROM aft_customer
             WHERE icnumber=" + TextBox1.Text + "", con)
-            con.Open()
-            command.ExecuteNonQuery()
-            MessageBox.Show("Customer Info Deleted")
-            con.Close()
-            TextBox1.Text = ""
-            GroupBox2.Hide()
-            GroupBox3.Hide()
-        Catch ex As Exception
-            MessageBox.Show("Exception: {0}", ex.Message)
-            MessageBox.Show("We found an error, Please contact with super admin.")
-        End Try
+                con.Open()
+                command.ExecuteNonQuery()
+                MessageBox.Show("Customer Info Deleted")
+                con.Close()
+                TextBox1.Text = ""
+                GroupBox2.Hide()
+                GroupBox3.Hide()
+            Catch ex As Exception
+                MessageBox.Show("Exception: {0}", ex.Message)
+                MessageBox.Show("We found an error, Please contact with super admin.")
+            End Try
+        Else
+            MessageBox.Show("Please Enter IC Number.")
+        End If
     End Sub
 End Class
